requirements.txt
celery==4.4.7
redis==3.5.3
Faker==6.6.0
python-dotenv==0.15.0
flower==0.9.7
newrelic==6.2.0.156
boto3==1.17.30
botocore==1.20.30
gevent==21.1.2